# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "sakalWeb Information System v5 RESTful API documentation",
        "title": "swis-api v5",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://savla.dev/swapi",
            "email": "info@savla.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/savla-dev/swis-api/blob/master/LICENSE"
        },
        "version": "5.2.4"
    },
    "host": "swapi.savla.su:8049",
    "basePath": "/",
    "paths": {
...
        "/backups/status/{service}": {
            "get": {
                "description": "get backup status by project/service",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backups"
                ],
                "summary": "Get backup status by project/service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dish instance name",
                        "name": "host",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/backups/{service}": {
            "put": {
                "description": "update backup status by service",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backups"
                ],
                "summary": "Update backup status by service",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backups.Backup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backups.Backup"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete backup service by its Name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backups"
                ],
                "summary": "Delete backup service by its Name",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/backups/{service}/active": {
            "put": {
                "description": "active/inactive backup toggle by its ServiceName",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backups"
                ],
                "summary": "Acitive/inactive backup toggle by its ServiceName",
                "parameters": [
                    {
                        "type": "string",
                        "description": "service name",
                        "name": "service_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backups.Backup"
                        }
                    }
                }
            }
        },
...
        "/dish/sockets/{host}": {
            "get": {
                "description": "get socket list by Host",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dish"
                ],
                "summary": "Get socket list by host",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dish instance name",
                        "name": "host",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/dish/sockets/{id}": {
            "put": {
                "description": "update socket by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dish"
                ],
                "summary": "Update socket by its ID",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dish.Socket"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dish.Socket"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete socket by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dish"
                ],
                "summary": "Delete socket by its ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dish ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dish.Socket"
                        }
                    }
                }
            }
        },
        "/dish/sockets/{id}/mute": {
            "put": {
                "description": "mute/unmute socket by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dish"
                ],
                "summary": "Mute/unmute socket by its ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dish ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dish.Socket"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "backups.Backup": {
            "type": "object",
            "required": [
                "service_name"
            ],
            "properties": {
                "active": {
                    "description": "Boolean indicating if the service is to be backuped.",
                    "type": "boolean"
                },
                "backup_size": {
                    "description": "Size of the gzip/tar archive.",
                    "type": "string"
                },
                "description": {
                    "description": "More verbose description of such service backup.",
                    "type": "string"
                },
                "executor_url": {
                    "description": "Dumping script, git URL.",
                    "type": "string"
                },
                "file_name": {
                    "description": "Name of the compressed backup file.",
                    "type": "string"
                },
                "last_status": {
                    "description": "Last status string of such backup (e.g. success, failure).",
                    "type": "string",
                    "default": "unknown"
                },
                "service_name": {
                    "description": "Backuped service name -- unique identifier.",
                    "type": "string"
                },
                "timestamp": {
                    "description": "UNIX timestamp of the last provided backup.",
                    "type": "integer"
                }
            }
        },
...
        "dish.Socket": {
            "type": "object",
            "required": [
                "host_name",
                "port_tcp",
                "socket_id"
            ],
            "properties": {
                "dish_target": {
                    "description": "DishTarget is a string array, usually containing dish's host short name (e.g. frank)\nto be refered e.g. /dish/sockets/frank",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "expected_http_code_array": {
                    "description": "If the Host is HTTP/S endpoint, one can specify which HTTP Result/Response codes are okay and not to alert upon",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "host_name": {
                    "description": "Hostname (server.random.com) or HTTP/S URI (http://endpoint.space)",
                    "type": "string"
                },
                "muted": {
                    "description": "Muted bool indicates that the socket is not propagated to any dish",
                    "type": "boolean"
                },
                "path_http": {
                    "description": "PathHTTP is any URL the site is about to be tested on, e.g. /dish/sockets",
                    "type": "string"
                },
                "port_tcp": {
                    "description": "Socket TCP port part\nEven default port 80 should be added here",
                    "type": "integer"
                },
                "socket_description": {
                    "description": "More verbose name/description of the socket",
                    "type": "string"
                },
                "socket_id": {
                    "description": "Socket ID, snake_cased for socket editing and deleting",
                    "type": "string"
                },
                "socket_name": {
                    "description": "GEneric name of the socket, to be used in dish results as failed one endpoint for example",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "apiKey": {
            "type": "apiKey",
            "name": "X-Auth-Token",
            "in": "header"
        }
    }
}